Erasing Phenotype Stuff
Since this branch will focus on genotype based crossing, which is more important than with phenotypes, it will be more clutter-free and be easier for me to work with.

I don't know if I'l even add in support for using phenotypes.
I'l definitely do genotypes, maybe phenotypes later when everything's finished. Idk.