Basic Table
I made a very basic punnet square as a table in the output.
It doesn't have lines or anything, but the data is lined up correctly.

All I have to do now is before storing the combined alleles in the offspring variables,  detect which allele variable is capital (if one only) and place that allele first. This will make results like "lL" become "Ll," which is how it should be correctly written.