Upgraded to Python 3 + VS Studio and Adding Genotype Percents
SO

I wanted to get this program on Python 3, so I did that. Turns out
Enthought Canopy doesn't work with anything above Python 2.X, so I
switched over to Visual Studio and imported my stuff and made a new
project there, so VS made some files.

Next, I started work on getting the percents and fractions of the unique
genotypes in the final punnet square