Display Genotype Percents
Yup, now in the final printout of the punnet square, the individual
offspring genotypes and their corresponding percents are displayed as
well.

I think in my next commit I'll remove everything pertaining to cats -
though they're cool and my project did start off about the genetics of
cats, it's really just a punnet square machine now, so renaming the
project and stuff in the code to be organism-neutral makes more sense.
\o/